Calculate length of ASF files with more than 1 second precision. Based on patch by On2.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@29881 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/asf.h
@@ -222,7 +222,7 @@ struct asf_priv {
     int scrambling_b;
     unsigned packetsize;
     double   packetrate;
-    unsigned movielength;
+    double movielength;
     int asf_is_dvr_ms;
     uint32_t asf_frame_state;
     int asf_frame_start_found;
